Shader Optimization
Tested. OK

Added some shader optimization.
Added the teacher version. Array are working but transferring array from
C++ prog and the shader doesn't and we don't have the time to debbug...
(ShaderParameter gives a pointer instead of an array and the Shader expect
an array but receive a simple pointer instead.)